Support check response content (#5)
* Support check response content

Use the following to check for content:

```yaml
https://docs.readthedocs.io/en/stable/:
  content:
    - '<sript src="/_/static/javascript/readthedocs-analytics.js"></script>'
    - '<meta property="og:title" content="Read the Docs: documentation simplified" />'
```

The response has to be return successfully to perform the check, otherwise, it's
ignored. There are some improvements we could add here, but I think it's a good
start for our use case: "check if the addons has been injected"

Related https://github.com/readthedocs/readthedocs-ops/issues/1442

* Update the spec to support `content.present` and `content.absent`

Now, we can check if the content is or isn't in the response.

* Better error message

* Remove check for `response.ok`